MAINT: Implement builder as nested class
