+++ src/apps/cc/app.cpp
@@ -43,7 +43,6 @@
 #include "config/exception.h"
 #include "config/validators.h"
 #include "detector/arrival.h"
-#include "detector/detectorbuilder.h"
 #include "eventstore.h"
 #include "log.h"
 #include "magnitude/regression.h"
